chore: remove a debug output that has been left in
