+++ plugins/gatsby-transformer-nz-author/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "gatsby-transformer-nz-author",
+  "version": "1.0.0",
+  "private": true
+}
