+++ plugins/gatsby-transformer-nz-blog/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "gatsby-transformer-nz-blog",
+  "version": "1.0.0",
+  "private": true
+}
