--- plugins/gatsby-transformer-source-split/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "gatsby-transformer-source-split",
-  "version": "1.0.0",
-  "private": true
-}
